requirements.txt
django==4.2.5
graphene-django==3.1.5
django-filter==23.3
django-graphql-jwt==0.4.0
django-cors-headers==4.2.0
python-dotenv==1.0.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pytest-django
python-decouple
openpyxl